[FIXED] type应该用uint8存
Former-commit-id: 86cc68378da4c7fb82c884358a03ad6b553e0bb4

+++ coceditor/src/libcoco/AnimToBin.cpp
@@ -10,7 +10,7 @@ int AnimToBin::Size(const PackAnimation* anim)
 	int sz = 0;
 
 	sz += sizeof(uint16_t);				// id
-	sz += sizeof(uint16_t);				// type
+	sz += sizeof(uint8_t);				// type
 
 	sz += sizeof(uint16_t);				// components size
 	for (int i = 0, n = anim->components.size(); i < n; ++i) {
@@ -37,7 +37,7 @@ void AnimToBin::Pack(const PackAnimation* anim, uint8_t** ptr)
 	uint16_t id = anim->GetID();
 	pack(id, ptr);
 
-	uint16_t type = TYPE_ANIMATION;
+	uint8_t type = TYPE_ANIMATION;
 	pack(type, ptr);
 
 	// components
